+++ app/database/password_reset.py
@@ -7,7 +7,7 @@ from datetime import datetime
 
 from app.helpers import utcnow
 
-from sqlalchemy import BigInteger, Column, ForeignKey
+from sqlalchemy import Column, ForeignKey, Integer
 from sqlmodel import Field, SQLModel
 
 
@@ -17,7 +17,7 @@ class PasswordReset(SQLModel, table=True):
     __tablename__: str = "password_resets"
 
     id: int | None = Field(default=None, primary_key=True)
-    user_id: int = Field(sa_column=Column(BigInteger, ForeignKey("lazer_users.id"), nullable=False, index=True))
+    user_id: int = Field(sa_column=Column(Integer, ForeignKey("lazer_users.id"), nullable=False, index=True))
     email: str = Field(index=True)
     reset_code: str = Field(max_length=8)  # 8-character reset code
     created_at: datetime = Field(default_factory=utcnow)
